# data-collector/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 数据采集服务API文档，提供股票、财务、新闻和政策数据采集功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Data Collector API
  version: "1.0"
paths:
...
  /collect/stock/adj-factors:
    post:
      consumes:
      - application/json
      description: 采集指定日期的复权因子数据
      parameters:
      - description: 交易日期 (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: 股票代码列表，逗号分隔，为空则采集所有股票
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 按日期采集复权因子数据
      tags:
      - 复权因子采集
  /collect/stock/adj-factors/latest:
    post:
      consumes:
      - application/json
      description: 采集最新交易日的复权因子数据
      parameters:
      - description: 股票代码列表，逗号分隔，为空则采集所有股票
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 采集最新复权因子数据
      tags:
      - 复权因子采集
  /collect/stock/adj-factors/range:
    post:
      consumes:
      - application/json
      description: 采集指定日期范围的复权因子数据
      parameters:
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - description: 股票代码列表，逗号分隔，为空则采集所有股票
        in: query
        name: symbols
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 采集成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 按日期范围采集复权因子数据
      tags:
      - 复权因子采集
...
  /stocks/adj-factors/by-date:
    get:
      consumes:
      - application/json
      description: 查询指定日期的所有股票复权因子数据
      parameters:
      - description: 交易日期 (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      - description: 返回数量限制，默认100
        in: query
        name: limit
        type: integer
      - description: 偏移量，默认0
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 查询指定日期的复权因子数据
      tags:
      - 复权因子查询
  /stocks/adj-factors/by-symbol:
    get:
      consumes:
      - application/json
      description: 查询指定股票在指定时间范围内的复权因子数据
      parameters:
      - description: 股票代码
        in: query
        name: symbol
        required: true
        type: string
      - description: 开始日期 (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: 结束日期 (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 查询指定股票的复权因子数据
      tags:
      - 复权因子查询
schemes:
- http
- https
swagger: "2.0"